backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
python-telegram-bot==20.3
python-dotenv==1.0.0
gunicorn==20.1.0
django-cors-headers==4.0.0
celery==5.3.0
redis==4.5.4
reportlab==4.2.5
requests==2.31.0